// frontend/app/src/project/directusTypes/types.ts
export type Canteens = {
  building?: number & Buildings;
  date_created?: string;
  date_updated?: string;
  id: number;
  label?: string;
  sort?: number;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
  businesshours: string & CanteensBusinesshours[];
};
// ...
export type Devices = {
  brand?: string;
  date_created?: string;
  date_updated?: string;
  display_fontscale?: string;
  display_height?: string;
  display_pixelratio?: string;
  display_scale?: string;
  display_width?: string;
  id: number;
  isAndroid?: boolean;
  isIOS?: boolean;
  isLandscape?: boolean;
  isSimulator?: boolean;
  isTablet?: boolean;
  isWeb?: boolean;
  platform?: string;
  profile?: number & Profiles;
  pushTokenObj?: unknown;
  sort?: number;
  status: string;
  systemVersion?: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
  display_group: string;
  system_group: string;
};

export type Flowhooks = {
  date_created?: string;
  date_updated?: string;
  id: number;
  last_parsing_date?: string;
  parse_foodoffers?: string;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
};

export type Foodoffers = {
  canteen?: number & Canteens;
  date?: string;
  date_created?: string;
  date_updated?: string;
  food?: string & Foods;
  id: number;
  price_employee?: number;
  price_guest?: number;
  sort?: number;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
  price_student?: number;
  markings: string & FoodoffersMarkings[];
};

export type FoodoffersMarkings = {
  foodoffers_id?: number & Foodoffers;
  id: number;
  markings_id?: number & Markings;
};

export type Foods = {
  calories_kcal?: number;
  carbohydrate_g?: number;
  date_created?: string;
  date_updated?: string;
  extra?: unknown;
  fat_g?: number;
  fiber_g?: number;
  id: string;
  image?: string & DirectusFiles;
  protein_g?: number;
  saturated_fat_g?: number;
  sodium_g?: number;
  sort?: number;
  status: string;
  sugar_g?: number;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
  feedbacks: string & FoodsFeedbacks[];
  markings: string & FoodsMarkings[];
  translations: string & FoodsTranslations[];
};

export type FoodsFeedbacks = {
  comment?: string;
  foods_id?: string & Foods;
  id: number;
  profiles_id?: number & Profiles;
  rating?: number;
};

export type FoodsMarkings = {
  foods_id?: string & Foods;
  id: number;
  markings_id?: number & Markings;
};

export type FoodsTranslations = {
  foods_id?: string & Foods;
  id: number;
  languages_code?: string & Languages;
  name?: string;
};
// ...
export type Languages = {
  code: string;
  direction?: string;
  name?: string;
};

export type Markings = {
  date_created?: string;
  date_updated?: string;
  icon?: string;
  id: number;
  image?: string & DirectusFiles;
  image_url?: string;
  label?: string;
  sort?: number;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
  translations: string & MarkingsTranslations[];
};

export type MarkingsTranslations = {
  be_source_for_translations?: boolean;
  create_translations_for_all_languages?: boolean;
  id: number;
  languages_code?: string & Languages;
  let_be_translated?: boolean;
  markings_id?: number & Markings;
  name?: string;
};
// ...
export type Profiles = {
  avatar?: unknown;
  canteen?: number & Canteens;
  course_timetable?: unknown;
  credit_balance?: number;
  date_created?: string;
  date_updated?: string;
  id: number;
  language?: string;
  nickname?: string;
  sort?: number;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
  buildings_favorites: string & ProfilesBuildingsFavorites[];
  buildings_last_visited: string & ProfilesBuildingsLastVisited[];
  devices: string & Devices[];
  foods_feedbacks: string & FoodsFeedbacks[];
  markings: string & ProfilesMarkings[];
};

export type ProfilesBuildingsFavorites = {
  buildings_id?: number & Buildings;
  id: number;
  profiles_id?: number & Profiles;
};

export type ProfilesBuildingsLastVisited = {
  buildings_id?: number & Buildings;
  id: number;
  last_visited?: string;
  profiles_id?: number & Profiles;
};

export type ProfilesMarkings = {
  dislikes?: boolean;
  id: number;
  markings_id?: number & Markings;
  profiles_id?: number & Profiles;
};
// ...
export type DirectusFiles = {
  id: string;
  storage: string;
  filename_disk?: string;
  filename_download: string;
  title?: string;
  type?: string;
  folder?: string & DirectusFolders;
  uploaded_by?: string & DirectusUsers;
  uploaded_on: string;
  modified_by?: string & DirectusUsers;
  modified_on: string;
  charset?: string;
  filesize?: number;
  width?: number;
  height?: number;
  duration?: number;
  embed?: string;
  description?: string;
  location?: string;
  tags?: unknown;
  metadata?: unknown;
  storage_divider: string;
};

export type DirectusFolders = {
  id: string;
  name: string;
  parent?: string & DirectusFolders;
};
// ...
export type DirectusRoles = {
  id: string;
  name: string;
  icon: string;
  description?: string;
  ip_access?: unknown;
  enforce_tfa: boolean;
  admin_access: boolean;
  app_access: boolean;
  users: string & DirectusUsers[];
};
// ...
export type DirectusUsers = {
  id: string;
  first_name?: string;
  last_name?: string;
  email?: string;
  password?: string;
  location?: string;
  title?: string;
  description?: string;
  tags?: unknown;
  avatar?: string & DirectusFiles;
  language?: string;
  theme?: string;
  tfa_secret?: string;
  status: string;
  role?: string & DirectusRoles;
  token?: string;
  last_access?: string;
  last_page?: string;
  provider: string;
  external_identifier?: string;
  auth_data?: unknown;
  email_notifications?: boolean;
  profile?: number & Profiles;
  preferences_divider: string;
  admin_divider: string;
};